Добавил CRUD систему, связанную с сущностями VerificationToken, ResetPasswordToken.
 Создал страницу с настройками, и соотв. функционалом: смена почты и пароля, удаление аккаунта, выход из аккаунта. Создал страницу подписчиков, также логику подписки и отписки от пользователя.

Создал страницу с поиском всех пользователей.
Обновил MathJax.
Описал все классы и методы в соотв. с JavaDoc форматом. Теперь рейтинг пользователя зависит от рейтинга всех его комментариев (суммарный ретйинг)

+++ src/main/java/mutata/com/github/MatematixProject/entity/dto/CommentDTO.java
@@ -15,7 +15,7 @@ public class CommentDTO {
     /**
      * Рейтинг комментария
      */
-    private long rating;
+    private int rating;
     /**
      * Дата, когда был оставлен комментарий
      */
